register_karo_agent/requirements.txt
# Server and API dependencies
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6
websockets>=11.0.3
jinja2>=3.1.2
python-dotenv>=1.0.0

# OpenAI API for document verification
openai>=1.0.0

# Razorpay API for payment processing
razorpay>=1.3.0

# Database (for future use)
sqlalchemy>=2.0.0

# Data validation and modeling
pydantic>=2.0.0